+++ src/testUtils/index.js
@@ -1,1 +1,3 @@
 export { default as mergeTestDescriptions } from "./mergeTestDescriptions"
+export { default as createRuleTester } from "./createRuleTester"
+export { default as testRule } from "./testRule"